Split styles by corresponding components
