+++ app/src/pages/inside/dashboardPage/widgets/components/testsTableWidget/testsTable/count/count.scss
@@ -1,0 +1,7 @@
+.count {
+  color: $COLOR--charcoal-grey;
+  position: absolute;
+  top: 50%;
+  transform: translateY(-50%);
+  left: 60px;
+}
