+++ app/src/pages/inside/dashboardPage/widgets/components/testsTableWidget/testsTable/count/index.jsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { number } from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 import classNames from 'classnames/bind';
-import styles from '../testsTableWidget.scss';
+import styles from './count.scss';
 
 const cx = classNames.bind(styles);
