+++ app/src/pages/inside/dashboardPage/widgets/components/testsTableWidget/testsTable/testsTable.scss
@@ -1,0 +1,4 @@
+.tests-table {
+  flex: 1;
+  padding-bottom: 20px;
+}
