+++ app/src/pages/inside/dashboardPage/widgets/components/testsTableWidget/testsTable/testsTableBody/index.jsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
 import { func } from 'prop-types';
 import classNames from 'classnames/bind';
-import { PTTests, PTColumns } from '../pTypes';
-import { TestsTableRow } from './testsTableRow';
-import { matrixFactory } from './matrix';
-import styles from '../testsTableWidget.scss';
+import { PTTests, PTColumns } from '../../pTypes';
+import { TestsTableRow } from '../testsTableRow';
+import { matrixFactory } from '../matrix';
+import styles from './testsTableBody.scss';
 
 const cx = classNames.bind(styles);
