+++ app/src/pages/inside/dashboardPage/widgets/components/testsTableWidget/testsTable/testsTableBody/testsTableBody.scss
@@ -1,0 +1,4 @@
+.tests-table-body {
+  width: 100%;
+  overflow: hidden;
+}
