+++ app/src/pages/inside/dashboardPage/widgets/components/testsTableWidget/testsTable/testsTableRow/testsTableRow.scss
@@ -1,0 +1,83 @@
+.row {
+  display: table;
+  table-layout: fixed;
+  margin-right: 0;
+  margin-left: 0;
+  height: 45px;
+  width: 100%;
+  border-bottom: 1px solid $COLOR--gray-91;
+  font-size: 13px;
+  &:hover {
+    background-color: $COLOR--tealish-hover;
+  }
+  &:last-child {
+    border-bottom: none;
+  }
+  .col {
+    display: table-cell;
+    vertical-align: middle;
+    padding-right: 5px;
+    &.col-name {
+      width: 40%;
+      font-size: 14px;
+      font-family: $FONT-SEMIBOLD;
+      color: $COLOR--windows-blue;
+      white-space: nowrap;
+      overflow: hidden;
+      text-overflow: ellipsis;
+      span {
+        cursor: pointer;
+      }
+    }
+    &.col-count {
+      width: 20%;
+      padding: 5px 5px 5px 0;
+      position: relative;
+      @media (max-width: $SCREEN_XS_MAX) {
+        width: 40%;
+      }
+    }
+    &.col-percents {
+      width: 20%;
+      color: $COLOR--charcoal-grey;
+      @media (max-width: $SCREEN_XS_MAX) {
+        width: 20%;
+      }
+    }
+    &.col-date {
+      width: 20%;
+      color: $COLOR--charcoal-grey;
+      cursor: pointer;
+      @media (max-width: $SCREEN_XS_MAX) {
+        display: none;
+      }
+      .relative {
+        display: none;
+      }
+      &:hover {
+        .absolute {
+          display: none;
+        }
+        .relative {
+          display: inline;
+        }
+      }
+      &.date-from-now {
+        .absolute {
+          display: none;
+        }
+        .relative {
+          display: inline;
+        }
+        &:hover {
+          .absolute {
+            display: inline;
+          }
+          .relative {
+            display: none;
+          }
+        }
+      }
+    }
+  }
+}
