+++ app/src/pages/inside/dashboardPage/widgets/components/testsTableWidget/testsTableWidget.scss
@@ -6,206 +6,5 @@
     padding: 0 20px 20px 20px;
     display: flex;
     flex-direction: column;
-    .info-block {
-      font-family: $FONT-SEMIBOLD;
-      font-size: 12px;
-      padding-bottom: 10px;
-      color: $COLOR--gray-47;
-      text-transform: uppercase;
-      white-space: nowrap;
-      text-overflow: ellipsis;
-      overflow: hidden;
-      p {
-        margin: 12px 0 0 0;
-        line-height: 17px;
-        .info-block-launch-name {
-          font-size: 13px;
-          color: $COLOR--charcoal-grey;
-          text-transform: none;
-        }
-      }
-    }
-    .tests-table {
-      flex: 1;
-      padding-bottom: 20px;
-      .tests-table-header {
-        height: 45px;
-        border-bottom: 1px solid $COLOR--gray-91;
-        background-color: $COLOR--white-two;
-        font-family: $FONT-SEMIBOLD;
-        font-size: 12px;
-        color: $COLOR--gray-47;
-        text-transform: uppercase;
-        .col {
-          float: left;
-          line-height: 45px;
-          white-space: nowrap;
-          overflow: hidden;
-          text-overflow: ellipsis;
-          &.col-name {
-            width: 40%;
-          }
-          &.col-count {
-            width: 20%;
-            .short {
-              display: none;
-            }
-            @media (max-width: $SCREEN_XS_MAX) {
-              width: 40%;
-              .full {
-                display: none;
-              }
-              .short {
-                display: inline;
-              }
-            }
-          }
-          &.col-percents {
-            width: 20%;
-            .short {
-              display: none;
-            }
-            @media (max-width: $SCREEN_XS_MAX) {
-              width: 20%;
-              .full {
-                display: none;
-              }
-              .short {
-                display: inline;
-              }
-            }
-          }
-          &.col-date {
-            width: 20%;
-            @media (max-width: $SCREEN_XS_MAX) {
-              display: none;
-            }
-          }
-        }
-      }
-      .tests-table-body {
-        width: 100%;
-        overflow: hidden;
-        .row {
-          display: table;
-          table-layout: fixed;
-          margin-right: 0;
-          margin-left: 0;
-          height: 45px;
-          width: 100%;
-          border-bottom: 1px solid $COLOR--gray-91;
-          font-size: 13px;
-          &:hover {
-            background-color: $COLOR--tealish-hover;
-          }
-          &:last-child {
-            border-bottom: none;
-          }
-          .col {
-            display: table-cell;
-            vertical-align: middle;
-            padding-right: 5px;
-            &.col-name {
-              width: 40%;
-              font-size: 14px;
-              font-family: $FONT-SEMIBOLD;
-              color: $COLOR--windows-blue;
-              white-space: nowrap;
-              overflow: hidden;
-              text-overflow: ellipsis;
-              span {
-                cursor: pointer;
-              }
-            }
-            &.col-count {
-              width: 20%;
-              padding: 5px 5px 5px 0;
-              position: relative;
-              @media (max-width: $SCREEN_XS_MAX) {
-                width: 40%;
-              }
-              .count {
-                color: $COLOR--charcoal-grey;
-                position: absolute;
-                top: 50%;
-                transform: translateY(-50%);
-                left: 60px;
-              }
-              .matrix {
-                display: table;
-                max-width: 50px;
-                min-height: 35px;
-                .squares-wrapper {
-                  display: table-cell;
-                  vertical-align: middle;
-                  .square {
-                    float: left;
-                    width: 4px;
-                    height: 4px;
-                    margin-right: 1px;
-                    margin-bottom: 1px;
-                    background-color: $COLOR--booger;
-                    &.most-failed {
-                      background-color: $COLOR--tealish;
-                    }
-                    &.passed {
-                      background-color: $COLOR--passed;
-                    }
-                    &.failed {
-                      background-color: $COLOR--failed;
-                    }
-                    &.skipped {
-                      background-color: $COLOR--skipped;
-                    }
-                  }
-                }
-              }
-            }
-            &.col-percents {
-              width: 20%;
-              color: $COLOR--charcoal-grey;
-              @media (max-width: $SCREEN_XS_MAX) {
-                width: 20%;
-              }
-            }
-            &.col-date {
-              width: 20%;
-              color: $COLOR--charcoal-grey;
-              cursor: pointer;
-              @media (max-width: $SCREEN_XS_MAX) {
-                display: none;
-              }
-              .relative {
-                display: none;
-              }
-              &:hover {
-                .absolute {
-                  display: none;
-                }
-                .relative {
-                  display: inline;
-                }
-              }
-              &.date-from-now {
-                .absolute {
-                  display: none;
-                }
-                .relative {
-                  display: inline;
-                }
-                &:hover {
-                  .absolute {
-                    display: inline;
-                  }
-                  .relative {
-                    display: none;
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    }
   }
 }
